switches time
